refactor(std/testing): Get rid of default export and make std/testing/diff.ts private (#7592)
